feat: merge auth layout
