+++ src/lib/types/database.types.ts
@@ -41,34 +41,37 @@ export interface Database {
 			Profile: {
 				Row: {
 					created_at: string | null;
-					cursus: string | null;
 					date_of_birth: string | null;
 					description: string | null;
 					first_name: string | null;
+					grade: string | null;
 					id: number;
 					last_name: string | null;
+					speciality: string | null;
 					status: boolean | null;
 					user_id: string | null;
 				};
 				Insert: {
 					created_at?: string | null;
-					cursus?: string | null;
 					date_of_birth?: string | null;
 					description?: string | null;
 					first_name?: string | null;
+					grade?: string | null;
 					id?: number;
 					last_name?: string | null;
+					speciality?: string | null;
 					status?: boolean | null;
 					user_id?: string | null;
 				};
 				Update: {
 					created_at?: string | null;
-					cursus?: string | null;
 					date_of_birth?: string | null;
 					description?: string | null;
 					first_name?: string | null;
+					grade?: string | null;
 					id?: number;
 					last_name?: string | null;
+					speciality?: string | null;
 					status?: boolean | null;
 					user_id?: string | null;
 				};
