+++ src/routes/(auth)/inscription/+layout.ts
